Split credentials into multiple components
